minor fixes and converted long to long long for MSVC

+++ week_1/day_6/day_6.cpp
@@ -11,7 +11,7 @@ int main(){
     std::vector<int> input = input_to_int(read_input("input_6", {","}));
 
     // vector of no of fish at each stage (index)
-    std::vector<long> stages(9);
+    std::vector<long long> stages(9);
 
     // fill school from input
     for (const auto& fish_stage : input){
@@ -20,7 +20,7 @@ int main(){
 
     // simulate school evolution for 80 days
     const int days = 256;
-    long part_1, part_2;
+    long long part_1, part_2;
     auto it = stages.begin();
     for (int i=0; i<days; i++){
 
@@ -33,11 +33,11 @@ int main(){
         if (++it == stages.end()){ it = stages.begin(); }
            
         // at 80 days record number of fish for part_1
-        if (i == 79){ part_1 = std::accumulate(stages.begin(),stages.end(),0l); }
+        if (i == 79){ part_1 = std::accumulate(stages.begin(),stages.end(),0ll); }
     }
 
     // count all fish 
-    part_2 = std::accumulate(stages.begin(),stages.end(),0l);
+    part_2 = std::accumulate(stages.begin(),stages.end(),0ll);
 
     std::cout << "Answer (part 1): " << part_1 << std::endl;
     std::cout << "Answer (part 2): " << part_2 << std::endl;
